Refactor: Use const for URL_PATTERNS
The URL_PATTERNS object is not intended to be modified. Changing it to a
`const` will help prevent accidental mutations and improve code clarity.
Refactor: Use const for URL_PATTERNS

The `URL_PATTERNS` object was defined using `let`, but it is never
reassigned. Changing it to `const` improves code clarity and indicates
that the object's reference will not change.